# api/openapi.yaml
openapi: 3.0.3
info:
  title: Zeno Auth API
  description: |
    Core authentication and identity service for the ZenoN-Cloud platform.
    
    This service provides:
    - User registration and authentication
    - JWT token management (access & refresh)
    - Organization management
    - GDPR compliance features
    - Session management
    
  version: 1.1.0
  contact:
    name: ZenoN-Cloud Team
    url: https://github.com/ZenoN-Cloud/zeno-auth
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://auth-dev.zenon-cloud.com
    description: Development environment
  - url: https://auth.zenon-cloud.com
    description: Production environment

tags:
  - name: Authentication
    description: User authentication endpoints
  - name: User
    description: User profile management
  - name: GDPR
    description: GDPR compliance endpoints
  - name: Sessions
    description: Session management
  - name: Consent
    description: User consent management
  - name: Health
    description: Service health checks
  - name: Admin
    description: Administrative endpoints

paths:
  /health:
    get:
      tags: [Health]
      summary: Basic health check
      description: Simple health check that always returns 200 OK
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: healthy
                  service:
                    type: string
                    example: zeno-auth

  /health/ready:
    get:
      tags: [Health]
      summary: Readiness probe
      description: Checks if service is ready to accept traffic (includes dependency checks)
      responses:
        '200':
          description: Service is ready
        '503':
          description: Service is not ready

  /health/live:
    get:
      tags: [Health]
      summary: Liveness probe
      description: Checks if service process is alive
      responses:
        '200':
          description: Service is alive

  /metrics:
    get:
      tags: [Health]
      summary: Prometheus metrics
      description: Returns service metrics in JSON format
      responses:
        '200':
          description: Metrics data
          content:
            application/json:
              schema:
                type: object

  /.well-known/jwks.json:
    get:
      tags: [Authentication]
      summary: JSON Web Key Set
      description: Returns public keys for JWT verification (RFC 7517)
      responses:
        '200':
          description: JWKS response
          content:
            application/json:
              schema:
                type: object
                properties:
                  keys:
                    type: array
                    items:
                      type: object

  /v1/auth/register:
    post:
      tags: [Authentication]
      summary: Register new user
      description: Creates a new user account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, full_name]
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                password:
                  type: string
                  format: password
                  minLength: 8
                  description: |
                    Password must meet the following requirements:
                    - At least 8 characters long
                    - At least 1 uppercase letter (A-Z)
                    - At least 1 lowercase letter (a-z)
                    - At least 1 digit (0-9)
                    - Must not be a common password
                  example: SecurePass123
                full_name:
                  type: string
                  example: John Doe
      responses:
        '201':
          description: User created successfully
        '400':
          description: Invalid input
        '409':
          description: Email already exists
        '429':
          description: Rate limit exceeded

  /v1/auth/login:
    post:
      tags: [Authentication]
      summary: User login
      description: Authenticates user and returns JWT tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                  refresh_token:
                    type: string
        '401':
          description: Invalid credentials
        '429':
          description: Rate limit exceeded

  /v1/auth/refresh:
    post:
      tags: [Authentication]
      summary: Refresh access token
      description: Generates new access token using refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token:
                  type: string
                  description: Valid refresh token
                  example: eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9...
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                    description: New JWT access token
                    example: eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9...
                  refresh_token:
                    type: string
                    description: New refresh token
                    example: eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9...
        '401':
          description: Invalid refresh token

  /v1/auth/logout:
    post:
      tags: [Authentication]
      summary: User logout
      description: Revokes all user's refresh tokens
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Logged out successfully
        '401':
          description: Unauthorized

  /v1/auth/verify-email:
    post:
      tags: [Authentication]
      summary: Verify email address
      description: Verifies user's email address using token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token:
                  type: string
      responses:
        '200':
          description: Email verified
        '400':
          description: Invalid token

  /v1/auth/resend-verification:
    post:
      tags: [Authentication]
      summary: Resend verification email
      description: Resends email verification link
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Verification email sent
        '401':
          description: Unauthorized

  /v1/auth/forgot-password:
    post:
      tags: [Authentication]
      summary: Request password reset
      description: Sends password reset link to user's email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: Reset link sent (if email exists)
        '429':
          description: Rate limit exceeded

  /v1/auth/reset-password:
    post:
      tags: [Authentication]
      summary: Reset password
      description: Resets password using reset token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, new_password]
              properties:
                token:
                  type: string
                new_password:
                  type: string
                  format: password
                  minLength: 8
      responses:
        '200':
          description: Password reset successful
        '400':
          description: Invalid token

  /v1/me:
    get:
      tags: [User]
      summary: Get user profile
      description: Returns current user's profile information
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User profile
        '401':
          description: Unauthorized

  /v1/me/change-password:
    post:
      tags: [User]
      summary: Change password
      description: Changes user's password
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [current_password, new_password]
              properties:
                current_password:
                  type: string
                  format: password
                new_password:
                  type: string
                  format: password
                  minLength: 8
      responses:
        '200':
          description: Password changed successfully
        '400':
          description: Invalid current password
        '401':
          description: Unauthorized
        '422':
          description: Validation error
        '500':
          description: Internal server error

  /v1/me/data-export:
    get:
      tags: [GDPR]
      summary: Export user data
      description: Returns all user data (GDPR Art. 15 - Right to Access)
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User data export
        '401':
          description: Unauthorized

  /v1/me/account:
    delete:
      tags: [GDPR]
      summary: Delete user account
      description: Deletes user account (GDPR Art. 17 - Right to be Forgotten)
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Account deleted
        '401':
          description: Unauthorized

  /v1/me/sessions:
    get:
      tags: [Sessions]
      summary: List active sessions
      description: Returns list of user's active sessions
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of sessions
        '401':
          description: Unauthorized
    delete:
      tags: [Sessions]
      summary: Revoke all sessions
      description: Revokes all user's sessions except current
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Sessions revoked
        '401':
          description: Unauthorized
        '500':
          description: Internal server error

  /v1/me/sessions/{id}:
    delete:
      tags: [Sessions]
      summary: Revoke specific session
      description: Revokes a specific session by ID
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Session revoked
        '400':
          description: Invalid session ID format
        '401':
          description: Unauthorized
        '404':
          description: Session not found
        '500':
          description: Internal server error

  /v1/me/consents:
    get:
      tags: [Consent]
      summary: Get user consents
      description: Returns list of user's consents
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of consents
        '401':
          description: Unauthorized
    post:
      tags: [Consent]
      summary: Grant consent
      description: Grants a specific consent
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [consent_type, version]
              properties:
                consent_type:
                  type: string
                  enum: [terms, privacy, marketing, analytics]
                version:
                  type: string
      responses:
        '200':
          description: Consent granted
        '400':
          description: Invalid consent type or version
        '401':
          description: Unauthorized
        '500':
          description: Internal server error

  /admin/compliance/report:
    get:
      tags: [Admin]
      summary: Get compliance report
      description: Returns GDPR compliance report
      security:
        - BearerAuth: []
      parameters:
        - name: start_date
          in: query
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Compliance report
        '401':
          description: Unauthorized
        '403':
          description: Forbidden (admin only)

  /admin/compliance/status:
    get:
      tags: [Admin]
      summary: Get compliance status
      description: Returns current GDPR and security compliance status
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Compliance status
        '401':
          description: Unauthorized
        '403':
          description: Forbidden (admin only)

  # Legacy endpoints (without /v1 prefix) for backward compatibility
  /auth/register:
    post:
      tags: [Authentication]
      summary: Register new user (legacy)
      deprecated: true
      description: |
        **DEPRECATED:** Use `/v1/auth/register` instead.
        This endpoint is maintained for backward compatibility.
      requestBody:
        $ref: '#/paths/~1v1~1auth~1register/post/requestBody'
      responses:
        $ref: '#/paths/~1v1~1auth~1register/post/responses'

  /auth/login:
    post:
      tags: [Authentication]
      summary: User login (legacy)
      deprecated: true
      description: |
        **DEPRECATED:** Use `/v1/auth/login` instead.
        This endpoint is maintained for backward compatibility.
      requestBody:
        $ref: '#/paths/~1v1~1auth~1login/post/requestBody'
      responses:
        $ref: '#/paths/~1v1~1auth~1login/post/responses'

  /auth/refresh:
    post:
      tags: [Authentication]
      summary: Refresh access token (legacy)
      deprecated: true
      description: |
        **DEPRECATED:** Use `/v1/auth/refresh` instead.
        This endpoint is maintained for backward compatibility.
      requestBody:
        $ref: '#/paths/~1v1~1auth~1refresh/post/requestBody'
      responses:
        $ref: '#/paths/~1v1~1auth~1refresh/post/responses'

  /auth/logout:
    post:
      tags: [Authentication]
      summary: User logout (legacy)
      deprecated: true
      description: |
        **DEPRECATED:** Use `/v1/auth/logout` instead.
        This endpoint is maintained for backward compatibility.
      security:
        - BearerAuth: []
      responses:
        $ref: '#/paths/~1v1~1auth~1logout/post/responses'

  /me:
    get:
      tags: [User]
      summary: Get user profile (legacy)
      deprecated: true
      description: |
        **DEPRECATED:** Use `/v1/me` instead.
        This endpoint is maintained for backward compatibility.
      security:
        - BearerAuth: []
      responses:
        $ref: '#/paths/~1v1~1me/get/responses'

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    Error:
      type: object
      properties:
        status:
          type: string
          example: error
        code:
          type: string
          example: invalid_input
        message:
          type: string
          example: Invalid input data

    SuccessResponse:
      type: object
      properties:
        status:
          type: string
          example: ok
        data:
          type: object

    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        full_name:
          type: string
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time

    Session:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_agent:
          type: string
        ip_address:
          type: string
        created_at:
          type: string
          format: date-time
        expires_at:
          type: string
          format: date-time

    Consent:
      type: object
      properties:
        consent_type:
          type: string
          enum: [terms, privacy, marketing, analytics, data_processing]
        version:
          type: string
        granted_at:
          type: string
          format: date-time
        revoked_at:
          type: string
          format: date-time
          nullable: true

# Global security: empty array means no default security requirement
# Individual endpoints override this with their own security requirements
security: []